requirements.txt
# Web Framework
fastapi>=0.100.0,<0.110.0
uvicorn[standard]>=0.20.0,<0.30.0

# LINE Bot SDK
line-bot-sdk>=3.0.0,<4.0.0

# Google Gemini AI
google-generativeai>=0.3.0,<1.0.0

# Data Processing
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
numpy>=1.21.0,<2.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0,<3.0.0

# HTTP Client
httpx>=0.24.0,<1.0.0
requests>=2.28.0,<3.0.0

# Utilities
python-multipart>=0.0.6

# Web Scraping for Movie Posters
beautifulsoup4>=4.12.0,<5.0.0
lxml>=4.9.0,<5.0.0
